+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.5.67:
 * Update `ReplicantSessionManagerImpl` so that `ChannelLinks` are only expanded for updates.
+* Update `ReplicantSessionManagerImpl` to make `delinkSubscriptionEntries` and `linkSubscriptionEntries`
+  protected access and available to subclasses.
 
 ## 0.5.66:
 * Update the `ReplicantSessionManagerImpl` so the ChangeSet is passed into many methods rather than