+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.5.68:
+* Introduce `AbstractSessionContextImpl` as a base class to extend as part of generation.
+
 ## 0.5.67:
 * Add `ReplicantSessionManagerImpl.delinkDownstreamSubscriptions` and exposed to subclasses.
 * Update `ReplicantSessionManagerImpl` so that `ChannelLinks` are only expanded for updates.